+++ stoq/plugins/connector.py
@@ -1,0 +1,12 @@
+from stoq.plugins.base import StoqPluginBase
+
+class StoqConnectorPlugin(StoqPluginBase):
+
+    def connect(self):
+        pass
+
+    def disconnect(self):
+        pass
+
+    def save(self):
+        pass
