+++ stoq/plugins/decorator.py
@@ -1,0 +1,7 @@
+from stoq.plugins.base import StoqPluginBase
+
+
+class StoqDecoratorPlugin(StoqPluginBase):
+
+    def decorate(self):
+        pass
