+++ stoq/plugins/extractor.py
@@ -1,0 +1,7 @@
+from stoq.plugins.base import StoqPluginBase
+
+
+class StoqExtractorPlugin(StoqPluginBase):
+
+    def extract(self):
+        pass
