+++ stoq/plugins/reader.py
@@ -1,0 +1,6 @@
+from stoq.plugins.base import StoqPluginBase
+
+class StoqReaderPlugin(StoqPluginBase):
+
+    def read(self):
+        pass
