+++ stoq/plugins/source.py
@@ -1,0 +1,8 @@
+import multiprocessing
+
+from stoq.plugins.base import StoqPluginBase
+
+class StoqSourcePlugin(StoqPluginBase, multiprocessing.Process):
+
+    def ingest(self):
+        pass
