+++ stoq/plugins/worker.py
@@ -1,103 +1,15 @@
-#   Copyright 2014-2018 PUNCH Cyber Analytics Group
-#
-#   Licensed under the Apache License, Version 2.0 (the "License");
-#   you may not use this file except in compliance with the License.
-#   You may obtain a copy of the License at
-#
-#       http://www.apache.org/licenses/LICENSE-2.0
-#
-#   Unless required by applicable law or agreed to in writing, software
-#   distributed under the License is distributed on an "AS IS" BASIS,
-#   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-#   See the License for the specific language governing permissions and
-#   limitations under the License.
-
-"""
-
-Overview
-========
-
-*StoqPluginManager()* is the primary class that controls all aspects of
-plugin management to include initialization, loading, listing, and unloading.
-This class is instantiated within the *Stoq()* class. This should not be
-instantiated outside of |stoQ| as it relies on objects within *Stoq()* to
-function properly.
-
-.. note:: Full plugin development documentation can be found at
-          :doc:`PluginDevelopment`.
-
-Examples
-========
-
-Instantiate Stoq::
-
-    from stoq.core import Stoq
-    stoq = Stoq()
-
-Listing all available plugins::
-
-    stoq.list_plugins()
-
-Once Stoq() is initialized, we can load a worker. The worker should always be
-instantiated first, then any additional plugins may be loaded through the
-worker plugin itself. The plugins will be instantiated within a dict in the
-worker plugin class. For example, a |stoQ| connector plugin may be accessed
-from it's plural name (connectors) within the worker object by calling
-``worker.connectors`` or a reader plugin may be called with
-``worker.readers``::
-
-    worker = stoq.load_plugin("yara", "worker")
-    worker.load_connector("file")
-    payload = worker.connectors['file'].get_file(path="/tmp/bad.exe")
-    results = worker.scan(payload)
-
-We may also retrieve a payload from a connector, such as MongoDB::
-
-    worker.load_connector("mongodb")
-    file_hash = "da39a3ee5e6b4b0d3255bfef95601890afd80709"
-    payload = worker.connectors['mongodb'].get_file(sha1=file_hash)
-    results = worker.scan(payload)
-
-.. note:: Only certain connector plugins support ``.get_file(**kwargs)``. Refer
-          to the plugin to determine if it is supported or not.
-
-Now that we have results, we can load our connector to save the results::
-
-   worker.connectors['mongodb'].save(results)
-
-We may also save a file via the connector. In this example, we will save a
-payload to with some additional attributes to GridFS::
-
-    payload_attributes = {}
-    payload_attributes['md5'] = "d41d8cd98f00b204e9800998ecf8427e"
-    payload_attributes['sha1'] = "da39a3ee5e6b4b0d3255bfef95601890afd80709"
-    payload_attributes['sha256'] = "e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855"
-    worker.connectors['mongodb'].save(payload, archive=True, payload_attributes)
-
-.. note:: ``save()`` accepts ``**kwargs``, so one may pass any attribute that
-          is needed to it. GridFS will automatically calculate the payload size
-          and datetime uploaded.
-
-API
-===
-"""
-
 import os
-import re
-import sys
-import glob
 import time
 import signal
-import logging
-import argparse
 import itertools
 import threading
-import subprocess
-import configparser
-import importlib.util
 import multiprocessing
 
-from pkg_resources import parse_version as version
+from stoq.exceptions import SigtermCaught
+from stoq import signal_handler, __version__
+from stoq.plugins.base import StoqPluginBase
+from stoq.helpers import ratelimited, flatten
+from stoq.scan import get_hashes, get_ssdeep, get_magic, get_sha1
 
 try:
     import yara
@@ -112,340 +24,6 @@ try:
     jinja_imported = True
 except ImportError:
     jinja_imported = False
-
-from stoq import signal_handler, __version__
-from stoq.helpers import ratelimited, flatten
-from stoq.exceptions import SigtermCaught
-from stoq.scan import get_hashes, get_ssdeep, get_magic, get_sha1
-
-
-class StoqPluginManager:
-    """
-
-    stoQ Plugin Manager Class
-
-    """
-
-    def __init__(self):
-
-        # Define the plugin categories and the associated class.
-        # If we need to add a new plugin category, it must be done here.
-        self.plugin_categories = {"worker": StoqWorkerPlugin,
-                                  "connector": StoqConnectorPlugin,
-                                  "reader": StoqReaderPlugin,
-                                  "source": StoqSourcePlugin,
-                                  "extractor": StoqExtractorPlugin,
-                                  "carver": StoqCarverPlugin,
-                                  "decoder": StoqDecoderPlugin,
-                                  "decorator": StoqDecoratorPlugin
-                                  }
-
-        self.plugin_extension = ".stoq"
-        self.collect_plugins()
-
-    @property
-    def __plugindict__(self):
-        """
-
-        Create a dict of plugin class name and the plugin category
-
-        """
-        plugins = {}
-        for category, category_class in self.plugin_categories.items():
-            plugins[category_class.__name__] = category
-
-        return plugins
-
-    def collect_plugins(self):
-        """
-        Find all stoQ plugins and their configuration file
-
-        """
-
-        self.__collected_plugins__ = {}
-
-        for plugin_dir_candidate in self.plugin_dir_list:
-            abs_plugin_path = os.path.abspath(plugin_dir_candidate.strip())
-            if not os.path.isdir(abs_plugin_path):
-                self.log.warning("Invalid plugin directory specified, "
-                              "skipping: {}".format(abs_plugin_path))
-                return
-
-            for root_path, subdirs, files in os.walk(abs_plugin_path):
-                for plg in files:
-                    if plg.endswith(self.plugin_extension):
-                        try:
-                            plugin_path = "{}/{}".format(root_path, plg)
-                            config = configparser.ConfigParser()
-                            config.read(plugin_path)
-                            name = config.get("Core", "Name")
-                            module = config.get("Core", "Module")
-                            module_path = "{}/{}.py".format(root_path, module)
-                            if os.path.isfile(module_path):
-                                # open each module file and detect the category of plugin
-                                with open(module_path, "r", encoding="utf-8") as src:
-                                    cat = re.search('(?<=from stoq\.plugins import )(.+)',
-                                                    src.read()).group()
-                                    category = self.__plugindict__.get(cat, False)
-
-                                config["Core"]["Category"] = category
-                                config["Core"]["Module"] = module_path
-                                config["Core"]["Root"] = plugin_dir_candidate
-                                self.__collected_plugins__[name] = config
-                            else:
-                                self.log.warning("Found {} but no module {}, skipping".format(plugin_path, module_path))
-                        except:
-                            self.log.error("Error parsing config file: {}".format(plugin_path))
-
-    @property
-    def get_categories(self):
-        """
-        Create list of plugin categories available
-
-        """
-
-        return self.plugin_categories.keys()
-
-    def get_plugins_of_category(self, category):
-        """
-        Lists plugin name of a specific category
-
-        :param str category: Category to discover plugins in
-
-        :returns: A tuple of discovered plugins
-        :rtype: tuple
-
-        """
-
-        for name, config in self.__collected_plugins__.items():
-            if config.get("Core", "Category") == category:
-                yield (config.get("Core", "Name"),
-                       config.get("Documentation", "Version"),
-                       config.get("Documentation", "Description")
-                       )
-
-    def get_plugin(self, name, category):
-        """
-        Initializes a plugin within a specific category
-
-        :param str name: Name of plugin to get
-        :param str category: Category of the named plugin
-
-        :returns: plugin object
-        :rtype: object
-
-        """
-
-        # Make sure we clean up argv so argument parsing of plugins
-        # doesn't cause issues when loading multiple worker plugins
-        if self.worker and self.argv:
-            self.argv = []
-
-        if category not in self.plugin_categories:
-            self.log.error("Invalid plugin category {}".format(category))
-            return False
-
-        info = self.__collected_plugins__.get(name)
-
-        if not info:
-            self.log.warning("No plugin available with the name {}".format(name))
-            return False
-
-        path = info.get("Core", "Module")
-        if not path:
-            self.log.error("No module found for {}".format(name))
-            return False
-
-        module_name = os.path.splitext(path)[0]
-
-        try:
-            spec = importlib.util.spec_from_file_location(module_name, path)
-            # Because module_from_spec wasn't introduced until python 3.5, we
-            # are going to check for it first. Let's hope this method doesn't
-            # get deprecated also...
-            if hasattr(importlib.util, 'module_from_spec'):
-                plugin = importlib.util.module_from_spec(spec)
-                spec.loader.exec_module(plugin)
-            else:
-                # Looks like we are at python 3.4, let's do it that way instead
-                plugin = spec.loader.load_module()
-
-            # Since the class we want to load for the plugin is a subclass
-            # of a plugin category, we are going to iterate over all classes
-            # in the module, and find the first one that is a subclass of
-            # plugin type we are looking for.
-            for plugin_object in dir(plugin):
-                try:
-                    if issubclass(getattr(plugin, plugin_object), self.plugin_categories[category]):
-                        if self.plugin_categories[category].__name__ is not plugin_object:
-                            loaded_plugin = getattr(plugin, plugin_object)
-                            loaded_plugin.details = info
-                            return loaded_plugin()
-                except TypeError:
-                    pass
-        except Exception:
-            self.log.error("Unable to load plugin", exc_info=True)
-            return False
-
-        return False
-
-    def load_plugin(self, name, category):
-        """
-        Load the desired plugin
-
-        :param str name: Plugin name to be loaded
-        :param str category: The category of plugin to be loaded
-
-        :returns: The loaded plugin object
-        :rtype: object
-
-        """
-
-        if not name or not category:
-            self.log.error("Attempted to load a plugin, but a name or category was not provided")
-            return None
-
-        # Initialize our plugin
-        self.log.debug("Attempting to load plugin {}:{}".format(category, name))
-        plugin = self.get_plugin(name, category)
-
-        if not plugin:
-            self.log.warning("Plugin {}:{} failed to load".format(category, name))
-            return False
-
-        for sect in plugin.details.sections():
-            for opt in plugin.details.options(sect):
-                # define each configuration option as an object within
-                # plugin class.
-                # Note: In order to reduce logic, we attempt to load
-                # the option as a boolean. By default, this will raise
-                # an error which in turn will cause us to load it as
-                # a string.
-                try:
-                    setattr(plugin, opt,
-                            plugin.details.getboolean(sect, opt))
-                except ValueError:
-                    value = plugin.details.get(sect, opt)
-                    if opt.endswith("_list"):
-                        # If our option ends with a list, let's turn it
-                        # into one
-                        # Example:
-                        # worker_list = this, is, a, list
-                        # Becomes:
-                        # worker.worker_list = ['this', 'is', 'a', 'list']
-                        value = [i.strip() for i in value.split(",")]
-                    elif opt.endswith("_dict"):
-                        value = self.loads(value)
-                    elif opt.endswith("_tuple"):
-                        value = tuple(i.strip() for i in value.split(","))
-                    elif opt.endswith("_int"):
-                        value = int(value.strip())
-
-                    setattr(plugin, opt, value)
-
-        setattr(plugin, 'category', category)
-        plugin_path = "{}/{}/{}".format(plugin.root, category, name)
-        plugin_path = os.path.abspath(plugin_path)
-        self.log.debug("{}:{} plugin path set to {}".format(category, name, plugin_path))
-        setattr(plugin, 'plugin_path', plugin_path)
-
-        # Make sure we attempt to activate the plugin after we setattr
-        # from the plugin config file
-        plugin.activate(self)
-        return plugin
-
-    @property
-    def get_all_plugin_names(self):
-        """
-        List all plugin names
-
-        :returns: All plugin names
-        :rtype: list
-
-        """
-
-        return self.__collected_plugins__.keys()
-
-    @property
-    def get_all_plugins(self):
-        """
-        List all valid plugins and configurations
-
-        :returns: All valid plugins
-        :rtype: dict
-
-        """
-
-        return self.__collected_plugins__
-
-    def list_plugins(self):
-        """
-        List all available plugins and their category
-
-        """
-
-        print("Available Plugins:")
-        for category in self.get_categories:
-            print(" {}s".format(category))
-            for plugin, ver, desc in self.get_plugins_of_category(category):
-                print("   - {}v{}{}".format(plugin.ljust(20),
-                                            str(ver).ljust(7),
-                                            desc))
-
-        return True
-
-class StoqPluginBase:
-
-    def __init__(self):
-        self.is_activated = False
-        self.min_stoq_version = None
-        self.max_stoq_version = None
-        self.incompatible_plugin = False
-        super().__init__()
-
-    @property
-    def min_version(self):
-        if self.min_stoq_version:
-            return version(__version__) >= version(self.min_stoq_version)
-        else:
-            return True
-
-    @property
-    def max_version(self):
-        if self.max_stoq_version:
-            return version(__version__) < version(self.max_stoq_version)
-        else:
-            return True
-
-    def activate(self):
-
-        # Instantiate the logging handler for this plugin
-        logname = "stoq.{}.{}".format(self.category, self.name)
-        self.log = logging.getLogger(logname)
-
-        if not self.min_version or not self.max_version:
-            self.incompatible_plugin = True
-            self.log.warning("Plugin not compatible with this version of stoQ. "
-                          "Unpredictable results may occur!")
-
-        # See if plugin options were provided when Stoq() was instantiated
-        plugin_options = self.stoq.plugin_options.get(self.category, {}).get(self.name, {})
-        for k in plugin_options:
-            if plugin_options[k] is not None:
-                setattr(self, k, plugin_options[k])
-
-        if hasattr(self, 'max_tlp'):
-            self.max_tlp = self.max_tlp.lower()
-
-        self.is_activated = True
-        self.log.debug("{} plugin activated".format(self.name))
-
-    def deactivate(self):
-        self.is_activated = False
-        self.log.debug("{} plugin deactivated".format(self.name))
-
-    def heartbeat(self, force=False):
-        pass
 
 
 class StoqWorkerPlugin(StoqPluginBase):
@@ -476,6 +54,7 @@ class StoqWorkerPlugin(StoqPluginBase):
         self.use_output_date = False
         self.flatten_results = False
         self.flatten_delimiter = False
+        self.ingest_metadata = None
 
         self.workers = {}
         self.connectors = {}
@@ -748,7 +327,7 @@ class StoqWorkerPlugin(StoqPluginBase):
 
         self.mp_queues.put(kwargs)
 
-    def scan(self):
+    def scan(self, *args, **kwargs):
         pass
 
     def load_worker(self, worker):
@@ -1230,7 +809,7 @@ class StoqWorkerPlugin(StoqPluginBase):
         # dispatched result, it will contain the plugin category as well as the
         # plugin name.
         if plugin.count(':') == 1:
-            plugin_cat, plugin_name = plugin.split(':')
+            _, plugin_name = plugin.split(':')
             index = self.name
         else:
             plugin_name = plugin
@@ -1396,270 +975,3 @@ class StoqWorkerPlugin(StoqPluginBase):
         if data['matches']:
             self.yara_dispatcher_hits.append(data)
         yara.CALLBACK_CONTINUE
-
-
-class StoqDecoratorPlugin(StoqPluginBase):
-
-    def decorate(self):
-        pass
-
-class StoqConnectorPlugin(StoqPluginBase):
-
-    def connect(self):
-        pass
-
-    def disconnect(self):
-        pass
-
-    def save(self):
-        pass
-
-
-class StoqReaderPlugin(StoqPluginBase):
-
-    def read(self):
-        pass
-
-
-class StoqSourcePlugin(StoqPluginBase, multiprocessing.Process):
-
-    def ingest(self):
-        pass
-
-
-class StoqExtractorPlugin(StoqPluginBase):
-
-    def extract(self):
-        pass
-
-
-class StoqCarverPlugin(StoqPluginBase):
-
-    def carve(self):
-        pass
-
-    def carve_payload(self, regex, payload, ignorecase=False):
-        """
-        Generator that returns a list of offsets for a specified value
-        within a payload
-
-        :param bytes regex: Regular expression to search for
-        :param bytes payload: Payload to be searched against
-        :param bool ignorecase: True or False, use re.IGNORECASE
-
-        :returns: Offset of value(s) discovered
-        :rtype: generator
-
-        """
-
-        self.log.debug("Carve: Attempting to carve payload")
-        try:
-            payload = payload.read()
-        except:
-            pass
-
-        if ignorecase:
-            # Ignorecase, Multiline, Dot matches all
-            flags = re.I|re.M|re.S
-        else:
-            # Multiline, Dot matches all
-            flags = re.M|re.S
-
-        for buff in re.finditer(regex, payload, flags):
-            self.log.debug("Carve: Payload carve at offset {} - {}".format(buff.start(), buff.end()))
-            yield buff.start(), buff.end()
-
-
-class StoqDecoderPlugin(StoqPluginBase):
-
-    def decode(self):
-        pass
-
-    def to_bytearray(self, payload):
-        """
-        Convert payload to a bytearray
-
-        :param bytes payload: Payload to be converted into byte array
-
-        :returns: Payload as a bytearray
-        :rtype: bytearray
-
-        """
-        self.log.debug("Converting payload ({} bytes) to a bytearray".format(len(payload)))
-        if isinstance(payload, bytearray):
-            pass
-        elif isinstance(payload, bytes):
-            payload = bytearray(payload)
-
-        else:
-            payload = bytearray(payload.encode())
-
-        return payload
-
-
-class StoqPluginInstaller:
-
-    pip_exists_str = "already exists. Specify --upgrade to force replacement."
-
-    def __init__(self, stoq):
-
-        self.stoq = stoq
-        self.plugin_info = {}
-
-        parser = argparse.ArgumentParser()
-        installer_opts = parser.add_argument_group("Plugin Installer Options")
-        installer_opts.add_argument("plugin", help="stoQ Plugin Archive")
-        installer_opts.add_argument("--upgrade",
-                                    action="store_true",
-                                    help="Upgrade the stoQ Plugin")
-        installer_opts.add_argument("-P", "--plugin-dir",
-                                    dest='plugin_dir',
-                                    default=False,
-                                    help="Root directory to install plugin to")
-
-        options = parser.parse_args(self.stoq.argv[2:])
-
-        if not options.plugin:
-            parser.print_help()
-            exit(-1)
-
-        # Set the source path of the plugin archive/directory
-        self.plugin = os.path.abspath(options.plugin)
-
-        # Define a directory to install a plugin to, if so desired
-        self.plugin_dir = options.plugin_dir
-
-        self.upgrade_plugin = options.upgrade
-
-    def install(self):
-        self.stoq.log.info("Looking for plugin in {}...".format(self.plugin))
-        try:
-            if os.path.isdir(self.plugin):
-                self.setup_from_dir()
-            else:
-                self.stoq.log.critical("Unable to install plugin. Is this a valid plugin?")
-                exit(-1)
-
-            try:
-                cmd = [
-                    sys.executable,
-                    '-m',
-                    'pip',
-                    'install',
-                    self.plugin,
-                    '-t',
-                    self.plugin_root,
-                ]
-                # Use pip to install/upgrade the plugin in the appropriate
-                # directory for this plugin category
-                if self.upgrade_plugin:
-                    cmd.append('--upgrade')
-
-                output = subprocess.check_output(cmd, stderr=subprocess.STDOUT)
-                if self.pip_exists_str.encode() in output:
-                    self.stoq.log.critical("Plugin {}".format(self.pip_exists_str))
-                    exit(-1)
-
-                # Time to install the requirements, if they exist.
-                requirements = "{}/requirements.txt".format(self.plugin)
-                if os.path.isfile(requirements):
-                    subprocess.check_call([
-                        sys.executable,
-                        '-m',
-                        'pip',
-                        'install',
-                        '--quiet',
-                        '-r',
-                        requirements,
-                    ])
-
-            except Exception as err:
-                self.stoq.log.critical("Error installing requirements: {}".format(str(err)))
-                exit(-1)
-
-        except FileNotFoundError as err:
-            self.stoq.log.critical(err)
-            exit(-1)
-
-        self.stoq.log.info("Install complete.")
-
-    def setup_from_dir(self):
-        # Find the stoQ configuration file
-        config_file = glob.glob("{}/*/*.stoq".format(self.plugin))
-
-        if len(config_file) > 1:
-            self.stoq.log.critical("More than one stoQ configuration file found. Exiting.")
-            exit(-1)
-
-        if os.path.isfile(config_file[0]):
-            # Open the stoQ configuration files and parse it
-            with open(config_file[0], "rb") as config_content:
-                self.parse_config(config_content.read())
-        else:
-            self.stoq.log.critical("Is this a valid configuration file? Exiting.")
-            exit(-1)
-
-        # Find the module name and set the plugin options
-        module_root = os.path.join(self.plugin, self.plugin_name)
-        module_path = os.path.join(module_root, self.plugin_module)
-        with open(module_path, "rb") as module_content:
-            self.set_plugin_category(module_content.read())
-
-        self.set_plugin_path()
-
-        self.save_plugin_info()
-
-        return True
-
-    def parse_config(self, stream):
-        config = configparser.ConfigParser()
-        config.read_string(stream.decode('utf-8'))
-        try:
-            self.plugin_name = config['Core']['Name']
-            self.plugin_module = "{}.py".format(config['Core']['Module'])
-
-            # We are going to use this to dynamically define data points in
-            # setup.py
-            self.plugin_info['NAME'] = self.plugin_name
-            self.plugin_info['AUTHOR'] = config['Documentation']['Author']
-            self.plugin_info['VERSION'] = config['Documentation']['Version']
-            self.plugin_info['WEBSITE'] = config['Documentation']['Website']
-            self.plugin_info['DESCRIPTION'] = config['Documentation']['Description']
-
-        except Exception as err:
-            self.stoq.log.critical("Is this a valid stoQ configuration file? {}".format(err))
-            exit(-1)
-
-    def save_plugin_info(self):
-        # Let's create text files with the appropriate attributes so setup.py
-        # can be more dynamic
-        for option, value in self.plugin_info.items():
-            with open(os.path.join(self.plugin, option), "w") as f:
-                f.write(value)
-
-    def set_plugin_category(self, plugin_stream):
-        # We've extract the StoqPlugin class that is specific to our plugin
-        # category, so now we can identity where the plugin will be
-        # installed into
-        try:
-            plugin_type = re.search('(?<=from stoq\.plugins import )(.+)',
-                                    plugin_stream.decode('utf-8')).group()
-            self.plugin_category = self.stoq.__plugindict__[plugin_type]
-        except Exception as err:
-            self.stoq.log.critical("Unable to determine the category. Is this a valid plugin? {}".format(err))
-            exit(-1)
-
-    def set_plugin_path(self):
-        if self.plugin_dir:
-            install_path  = self.plugin_dir
-        else:
-            if len(self.stoq.plugin_dir_list) > 1:
-                self.stoq.log.critical("Multiple plugin directories defined in stoq.cfg."
-                                       "Unable to determine plugin installation directory."
-                                       "Please explicitly define one using --plugin-dir")
-                exit(-1)
-            install_path = self.stoq.plugin_dir_list[0]
-
-        self.plugin_root = os.path.join(install_path, self.plugin_category)
-
-        self.stoq.log.info("Installing {} plugin into {}...".format(self.plugin_name, self.plugin_root))
